+++ lib/src/core/writers/assets_list_writer.dart
@@ -1,17 +1,23 @@
 import 'package:genuis/src/core/models/module.dart';
+import 'package:genuis/src/core/models/token.dart';
 
 class AssetsListWriter {
   AssetsListWriter();
 
-  void write(StringBuffer buffer, List<Module> modules) {
-    buffer.writeln('/*');
-    buffer.writeln('  assets:');
-    for (final module in modules) {
-      for (final asset in module.assetsList) {
+  void write(StringBuffer buffer, List<Module> modules, List<Token> tokens) {
+    final assets = {
+      ...tokens.expand((e) => e.assetsList),
+      ...modules.expand((e) => e.assetsList),
+    }.toList()..sort();
+
+    if (assets.isNotEmpty) {
+      buffer.writeln('/*');
+      buffer.writeln('  assets:');
+      for (final asset in assets) {
         buffer.writeln('    - $asset');
       }
+      buffer.writeln('*/');
+      buffer.writeln();
     }
-    buffer.writeln('*/');
-    buffer.writeln();
   }
 }