+++ test_packages/general_test/test/token_icons_test.dart
@@ -1,0 +1,33 @@
+import 'package:general_test/ui/ui.ui.dart';
+import 'package:test/test.dart';
+
+import 'values.dart';
+
+void main() {
+  test('Token: icons_tokens', () {
+    expect(IconsTokens.loremIpsumDolorAmet, testIcon01);
+    expect(IconsTokens.loremIpsumDolorSitDark, testIcon02);
+    expect(IconsTokens.loremIpsumDolorSitLight, testIcon03);
+    expect(IconsTokens.loremIpsumAdipiscing, testIcon04);
+    expect(IconsTokens.loremIpsumConsecteturDark, testIcon05);
+    expect(IconsTokens.loremIpsumConsecteturLight, testIcon06);
+    expect(IconsTokens.loremElitDark, testIcon07);
+    expect(IconsTokens.loremElitLight, testIcon08);
+    expect(IconsTokens.loremSed, testIcon09);
+    expect(IconsTokens.do_Dark, testIcon10);
+    expect(IconsTokens.do_Light, testIcon11);
+    expect(IconsTokens.eiusmod, testIcon12);
+  });
+}
+
+
+
+
+
+
+
+
+
+
+
+